+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 
 ## 2026-04-24
+- `18:12` -- **Updated** `apps/desktopProbe/src/lib/electronMainSdk.tsx`
 - `18:11` -- **Updated** `apps/desktopProbe/src/lib/electronMainSdk.tsx`
 - `18:11` -- **Updated** `apps/desktopProbe/src/server/rendererIpcApi.ts`
 - `18:10` -- **Updated** `apps/desktopProbe/src/server/jobScanner.ts`